simplify the code, find a bug, and write a test against a harder case
To any readers wondering about the commit date anomalies in this repo : Github has (had?) a visualization consisting of green squares that showed how many commits a given user had made. I was forging the dates because I wanted to have a commit for each day, and I wanted to have a buffer of commits so I don't actually need to commit every day.

If it is not clear why I wwrote this message in this commit, there was (is?) this thing called the "Github Artic Code Vault" that included a snapshot of every public github repo on 2020-02-02. I figured I would explain what might not be obvious to any readers of that archive.