NON-COMPILING: Try something but realize need a different design
Spend a while trying to make things faster, by avoiding stringifying ropes
by only stringifying the one I know that I'll need to stringify: the selected
one. But then the `SendBuffersToBeSaved` code/use case made me realize that the
better design is to allow each rope to be stringifed on demand. That way,
we will stringify less in practice, except when we do in fact need to
stringify everything.

(stringifying in this case implies making a copy of the rope and making the
memory contiguous so things like the std::fs API and rendering code know
how to work with it.)