NON-COMPILING: start implementing get_strip_trailing_whitespace_edit
But it seems like the API of tree sitter requires start and end (x,y) positions for each edit, and we ended up requiring each edit to have a cursor associated with it.
Not sure about the causality there.

But the cursors requirement does make `get_edit` make a bit more sense.