NON-COMPILING: start into a refactor, causing a chain reaction
Specifically, I think it makes sense to extract out a `History` struct, before we replace the `VecDeque` with a fixed sized buffer. But I also noticed we need to bundle up the Rope and Cursors to satisfy the borrow checker anyway, and also that we pass around a Rope and the Cursors a lot, so now I want to use the new struct everywhere it would have made sense to if we had started with it.